requirements.txt
Django==1.8.1
WebTest==2.0.18
coverage==3.7.1
dj-database-url==0.3.0
django-getenv==1.3.1
django-model-utils==2.2
django-webtest==1.7.8
djangorestframework==3.1.1
flake8==2.4.0
psycopg2==2.6
python-dotenv==0.1.2
uWSGI==2.0.10
